Move some utility functions from main.ts to new file utility.ts
